requirements.txt
Flask==1.0.2
qrcode==6.0
Pillow==5.2.0
